+++ src/config.ts
@@ -27,9 +27,11 @@ export type Config = {
     // API stuff
     backgroundLoadIntervalMS: number;
     defaultRecordsPerPage: number;
+    maxConcurrentTootRequests: number;
     maxRecordsForFeatureScoring: number;
     minRecordsForFeatureScoring: number;
     maxFollowingAccountsToPull: number;
+    mutexWarnSeconds: number;
     reloadFeaturesEveryNthOpen: number;
     staleDataSeconds: StaleDataConfig;
     timeoutMS: number;
@@ -96,8 +98,11 @@ export const DEFAULT_CONFIG: Config = {
     backgroundLoadIntervalMS: 60_000,    // 1 minute
     maxRecordsForFeatureScoring: 2_800,  // number of notifications, replies, etc. to pull slowly in background for scoring
     maxFollowingAccountsToPull: 5_000,   // MAX_FOLLOWING_ACCOUNT_TO_PULL
+    // TODO: possibly consider extending this to all API requests, not just hashtag toot pulls?
+    maxConcurrentTootRequests: 15,         // How many toot requests to make in parallel
     minRecordsForFeatureScoring: 240,    // number of notifications, replies, etc. to pull in initial load
     minServerMAU: 100,                   // Minimum MAU for a server to be considered for trending toots/tags
+    mutexWarnSeconds: 5,                 // How long to wait before warning about a mutex lock
     numServersToCheck: 30,               // NUM_SERVERS_TO_CHECK
     reloadFeaturesEveryNthOpen: 9,       // RELOAD_FEATURES_EVERY_NTH_OPEN
     timeoutMS: 5_000,                    // Timeout for API calls