+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Use a Semaphore to limit concurrency of hashtag search API requests; add Config.maxConcurrentTootRequests
 * Exclude followed tags from the pulling of hashtag participation toots (followed tags are already in timeline)
 * Refactor common handling for cacheable toot fetches